requirements.txt
altgraph==0.17.4
cffi==1.16.0
cryptography==42.0.8
packaging==24.1
pycparser==2.22
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pynput==1.7.7
python-dotenv==1.0.1
python-xlib==0.33
six==1.16.0
